requirements.txt
streamlit==1.38.0
pdfplumber==0.11.4
pydub==0.25.1
deepgram-sdk==3.7.2
TTS==0.22.0
torch==2.4.1
torchaudio==2.4.1
scipy==1.14.1
librosa==0.10.2
numpy==1.26.4